refactor: create paginated batch operations query
