+++ operate/client/src/modules/queries/batch-operations/useBatchOperations.ts
@@ -6,27 +6,16 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {useInfiniteQuery} from '@tanstack/react-query';
+import {useQuery} from '@tanstack/react-query';
 import {queryBatchOperations} from 'modules/api/v2/batchOperations/queryBatchOperations';
 import type {QueryBatchOperationsRequestBody} from '@camunda/camunda-api-zod-schemas/8.8';
-
-const BATCH_OPERATIONS_QUERY_KEY = 'batchOperations';
-const MAX_OPERATIONS_PER_REQUEST = 50;
+import {queryKeys} from '../queryKeys';
 
 const useBatchOperations = (payload: QueryBatchOperationsRequestBody) => {
-  return useInfiniteQuery({
-    queryKey: [BATCH_OPERATIONS_QUERY_KEY, payload],
-    queryFn: async ({pageParam}) => {
-      const requestPayload = {
-        ...payload,
-        page: {
-          ...payload.page,
-          limit: MAX_OPERATIONS_PER_REQUEST,
-          from: pageParam,
-        },
-      };
-
-      const {response, error} = await queryBatchOperations(requestPayload);
+  return useQuery({
+    queryKey: queryKeys.batchOperations.query(payload),
+    queryFn: async () => {
+      const {response, error} = await queryBatchOperations(payload);
 
       if (response !== null) {
         return response;
@@ -34,29 +23,7 @@ const useBatchOperations = (payload: QueryBatchOperationsRequestBody) => {
 
       throw error;
     },
-    initialPageParam: 0,
-    getNextPageParam: (lastPage, _, lastPageParam) => {
-      const {page} = lastPage;
-      const nextPage = lastPageParam + MAX_OPERATIONS_PER_REQUEST;
-
-      if (nextPage > page.totalItems) {
-        return null;
-      }
-
-      return nextPage;
-    },
-    getPreviousPageParam: (_, __, firstPageParam) => {
-      const previousPage = firstPageParam - MAX_OPERATIONS_PER_REQUEST;
-
-      if (previousPage < 0) {
-        return null;
-      }
-
-      return previousPage;
-    },
-    refetchInterval: 5000,
-    maxPages: 2,
   });
 };
 
-export {useBatchOperations, BATCH_OPERATIONS_QUERY_KEY};
+export {useBatchOperations};
