+++ operate/client/src/modules/queries/queryKeys.ts
@@ -9,6 +9,7 @@
 import type {
   ElementInstance,
   ProcessInstance,
+  QueryBatchOperationsRequestBody,
   QueryDecisionInstancesRequestBody,
   QueryElementInstanceIncidentsRequestBody,
   QueryElementInstancesRequestBody,
@@ -141,7 +142,10 @@ const queryKeys = {
     ],
   },
   batchOperations: {
-    get: (batchOperationKey: string) => ['batchOperation', batchOperationKey],
+    query: (payload: QueryBatchOperationsRequestBody) => [
+      'batchOperations',
+      payload,
+    ],
   },
   batchOperationItems: {
     searchByProcessInstanceKey: (processInstanceKey?: string) => [
